requirements.txt
pandas
slackclient
numpy
sqlalchemy
psycopg2
dropbox
tabulate
PyYAML
